requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-restframework==0.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
